refactor(import): PER-171 — harden Sure importer per head-eng review
Addresses the three non-negotiables from review + polish:

1. Real parity, not copy-vs-copy. Export isPromotable from the server
   (behavior-neutral; integration stays green) and assert the client preview
   classifier (sureHeldReason) returns an identical verdict across a full
   account×transaction branch matrix AND every row of both synthetic fixtures.
2. Preview fully reconciles to SureMigrationResult. summarizeSureBundle already
   tracks held / zeroAmount / invalidDate / unmapped; the review + done screens
   now surface every bucket (incl. unmapped, malformed lines, ignored entities)
   and a reconciliation strip shows total = importing + held + skipped, so no
   gap is unexplained. Unit test asserts the buckets sum to total.
3. Done screen is graceful at promotedThisRun=0 (the real degraded case): reads
   "Accounts & details imported … none were lost" instead of an empty/failed-
   looking "0". Extracted the presentational layer to -sure-import-ui.tsx and
   unit-tested the 0-promoted, promoted, and replayed states with RTL.

Polish already in place, reaffirmed: client 64 MiB guard + File.text() fail-fast;
bundle string lives only in state → server fn (no localStorage/log); Confirm
disabled + "Importing…" while in-flight, and the review screen is gone on success.
